Utility opgeschoond, java doc toegevoegd
